+++ CHANGELOG.markdown
@@ -1,10 +1,15 @@
 4.4
----
+next
 * Added `imapRep`, `ifoldMapRep`, `itraverseRep` to make it easier to define representable `FunctorWithIndex`, `FoldableWithIndex`, `TraversableWithIndex` instances from the `lens` package.
+* Add `GHC.Generics`-based default implementation for `Data.Functor.Rep.Representable` instances
+* Add `Data.Functor.Rep.Representable` instances for `Backwards`, `Reverse`, and the datatypes in `GHC.Generics`.
+* Add `Data.Functor.Adjunction.Adjunction` instances for some datatypes in `GHC.Generics`
+* Add `Data.Functor.Contravariant.Rep.Representable` instances for `U1` and `(:*:)` from `GHC.Generics`
+* Add `collectRep` and `imapRep` functions to `Data.Functor.Rep`.
+* Add `MINIMAL` pragmas to the `Adjunction` classes.
 
 4.3
 ---
-* Added a `MINIMAL` pragma to `Adjunction`.
 * Removed a spurious superclass constraint for `Applicative (StoreT g w)`
 * GHC 8 support
 * `comonad` 5 support